+++ changelog.md
@@ -2,6 +2,7 @@
 
 ### [Latest]
 
+- Update atlas-ftag-tools Version to v0.2.15 [#335](https://github.com/umami-hep/puma/pull/335)
 - Adding Support for VarVsEff Plots with Fixed Background Rejections [#334](https://github.com/umami-hep/puma/pull/334)
 - Updating Docstrings and skipped pre-commit checks [#332](https://github.com/umami-hep/puma/pull/332)
 - Update Repository Metadata [#331](https://github.com/umami-hep/puma/pull/331)